feat(tuic): wire brutal congestion control cwnd from share link
Parse/export the "cwnd" query param and thread it through Header.Feature2 into ClientOption.CWND; SetCongestionController selects brutal with cwnd as the target bandwidth (bytes/s, community convention), falling back to BBR when unset. Drops the dead CWND:10 hardcode in tuic and juicity.

+++ protocol/juicity/client.go
@@ -74,7 +74,7 @@ type ClientOption struct {
 	Uuid                 [16]byte
 	Password             string
 	CongestionController string
-	CWND                 int
+	CWND                 uint64
 	Ctx                  context.Context
 	Cancel               func()
 	UnderlayAuth         chan *UnderlayAuth
